rpc: partially allow param conversions for composite commands
The processing is done in RPCConvertValues() (RPC client) instead of
CRPCTable::execute() (RPC server) because we need to be able to populate
members(ByName) with arg positions without having to worry about the
offset-by-one as the parsing is done *before* execute() gets a chance
to account for the subcommand. Also because the parsing is done by the
client.

Named values are currently still not supported.