refactor: introduce `coinjoin status` for information on mix sessions
This helps us get rid of having to report it through `coinjoin start`.

Additionally, we won't try to jump ahead of the scheduler by starting
denomination routines ourselves, let the scheduler run its course.
This matters because we don't plan on exposing that capability through
the interface as we are migrating to the interface in the next commit.