net: reintroduce `IsSelectableSocket()` and make it SEM-aware
`IsSelectableSocket()` was subsumed into `Sock` for mockability's sake
but creates problems when our event-socket map uses the raw `SOCKET`
and not `Sock`, so we need to bring it back. On top of that,
`IsSelectableSocket()`'s behavior was defined at compile-time when it
should've been sensitive to the runtime capabilities we give it.

Currently, this does not cause a change in behavior but makes way for
future changes.