chore: remove scaffolding (remove default args, make explicit choice)
This was acceptable while we were refining earlier commits but default
args could in combination with future backports result in behavior
changes that were unanticipated, we should clear the air right now.

We are introducing a new global, `g_socket_events_mode` to allow us to
influence Wait(Many)() calls as and when needed. A global is acceptable
as it's lightweight, a parameter that is not configurable during runtime
except during initial startup and because the structure in which it is
ordinarily stored is far heavier.